TOPP: Changed interface of AdditiveSeries and SpectraFilter; ParamEditor: intoduced edit modes; Param: setValidStrings does not accept commas anymore; minor other changes
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@2741 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/VISUAL/MultiGradient.h
@@ -112,7 +112,7 @@ namespace OpenMS
 			OPENMS_PRECONDITION(pre_.size()!=0,"MultiGradient::precalculatedColorAt(DoubleReal): Precalculation mode not activated!");
 			OPENMS_PRECONDITION(position>=pre_min_,"MultiGradient::precalculatedColorAt(DoubleReal): Position out of specified range!");
 			OPENMS_PRECONDITION(position<=pre_min_+pre_size_,"MultiGradient::precalculatedColorAt(DoubleReal): Position out of specified range!");
-			return pre_[(position - pre_min_) / pre_size_ * pre_steps_];	
+			return pre_[(UInt)((position - pre_min_) / pre_size_ * pre_steps_)];	
 		}
 
 		///return the number of color points
